services/monitoring/requirements.txt
python-decouple
minio
python-decouple
mlflow
boto3